Add movie page refactored.
