+++ src/app/App.js
@@ -1,10 +1,10 @@
 import React from 'react';
 import {BrowserRouter, Route, Routes} from 'react-router-dom';
 
+import AddMovie from '../components/add_movie_screen/add_movie_screen.js';
 import MainScreen from '../components/main_screen/main_screen.js';
-import SearchModule from '../components/SearchMovieForm.js';
-import AddModule from '../components/AddMovieForm.js';
 import NavBar from '../components/navbar/navbar.js';
+import SearchModule from '../components/SearchMovieForm.js';
 
 function App() {
   return (
@@ -13,7 +13,7 @@ function App() {
         <Routes>
             <Route exact path = "/" element = {<MainScreen/>}/>
             <Route exact path = "/buscarpeliculas" element = {<SearchModule/>}/>
-            <Route exact path = "/agregarpeliculas" element = {<AddModule/>}/>
+            <Route exact path = "/agregarpeliculas" element = {<AddMovie/>}/>
             <Route exact path = "/modificarpeliculas" element = {<SearchModule/>}/>
             <Route exact path = "/eliminarpeliculas" element = {<SearchModule/>}/>
         </Routes>
