+++ src/components/search_movie_screen/search_movie_screen.css
